move EmbeddedResourceLoader into pathfinder_resources
