+++ gpu/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2018"
 
 [dependencies]
 half = "1.4"
-phf = "0.7"
 
 [dependencies.image]
 version = "0.21"
@@ -21,7 +20,3 @@ path = "../geometry"
 
 [dependencies.pathfinder_simd]
 path = "../simd"
-
-[build-dependencies]
-phf_codegen = "0.7"
-half = "1.4"
